+++ ParseWiki/ParseWiki/Sinks/NullSink.cs
@@ -4,7 +4,7 @@ namespace ParseWiki.Sinks
 {
     public class NullSink<T> : ISink<T>
     {
-        public async Task Save(int id, T item)
+        public async Task Save(long id, T item)
         {
         }
     }
